apps/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
PyJWT==2.8.0
pydantic==2.7.1
python-multipart==0.0.9
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.2
httpx==0.27.0

